requirements.txt
        tqdm
        prettytable
        torchinfo==1.8.0
        plyfile==0.9
        trimesh==4.1.8
        tensorboard
        pyvista
        rich==13.7.1
        CythonGSL
        loguru
        torchdata==0.7.1
        litdata
        transformers
        nltk
        ipykernel
        ipywidgets
        pillow
        accelerate
        python-dotenv
        gradio
        open3d
        scikit-learn
        timm
        wandb
        dgl